smite-scenarios: move tests to src/executor/tests.rs
Most of executor.rs was tests, which we now move to their own file.
